requirements.txt
anyio==3.6.2
autopep8==2.0.0
click==8.1.3
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
pycodestyle==2.10.0
pydantic==1.10.2
python-decouple==3.6
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
